Mostly Fix Tabs Issues
